@@ -1,5 +1,6 @@
 name: snow_weather_info
 description: A application to display french snow weather beacon informations
+publish_to: none
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -11,10 +12,10 @@ description: A application to display french snow weather beacon informations
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.2.2+10
+version: 1.6.0+16
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,27 +23,31 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  flutter_launcher_icons: "^0.7.3"
-  provider: ^4.3.2+1
-  http: ^0.12.0+4
-  latlong: ^0.6.1
-  csv: ^4.0.3
-  intl: ^0.16.1
-  flutter_map: any
-  charts_flutter: ^0.8.1
-  sqflite: ^1.3.1+1
-  shared_preferences: ^0.5.6+1
-  url_launcher: ^5.4.1
-  location: ^3.0.2
-  cached_network_image: ^2.5.0
-  flutter_cache_manager: ^2.1.0
-  package_info: ^0.4.3
-  native_pdf_view: ^3.8.0
-  carousel_slider: ^1.4.1
-  share: ^0.6.4+5
-  sticky_headers: ^0.1.8+1
-  alphabet_list_scroll_view: ^1.0.6
+  cupertino_icons: ^1.0.3
+  flutter_launcher_icons: "^0.9.0"
+  provider: ^5.0.0
+  http: ^0.13.3
+  collection: ^1.15.0
+  latlong2: ^0.8.0
+  csv: ^5.0.0
+  intl: ^0.17.0
+  flutter_map: ^0.13.1
+  # user_location: ^0.1.3
+  charts_flutter: ^0.11.0
+  sqflite: ^2.0.0+3
+  shared_preferences: ^2.0.6
+  url_launcher: ^6.0.6
+  package_info: ^2.0.2
+  carousel_slider: ^4.0.0-nullsafety.0
+  share: ^2.0.4
+  alphabet_list_scroll_view: ^2.0.0
+  webview_flutter: ^2.0.8
+  flutter_sticky_header: ^0.6.0
+  logging: ^1.0.1
+  dart_rss:
+    git:
+      url: https://github.com/Chonli/webfeed.git
+      ref: feature/geo-rss-2
 
 dev_dependencies:
   flutter_test:
@@ -52,6 +57,7 @@ flutter:
   uses-material-design: true
   assets:
     - assets/icon/icon.png
+    - assets/test_data/
 
 flutter_icons:
   android: "launcher_icon"
