@@ -1,15 +1,20 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:snow_weather_info/data/repository.dart';
-import 'package:snow_weather_info/ui/avalanche_massif_list_page.dart';
+import 'package:package_info/package_info.dart';
+import 'package:provider/provider.dart';
+import 'package:snow_weather_info/data/data_notifier.dart';
+import 'package:snow_weather_info/modules/brea/view.dart';
+import 'package:snow_weather_info/modules/avalanche/view.dart';
 import 'package:snow_weather_info/ui/list_station_widget.dart';
 import 'package:snow_weather_info/ui/map_widget.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:url_launcher/url_launcher.dart' as url;
 
 class HomePage extends StatefulWidget {
+  const HomePage({
+    Key? key,
+    required this.title,
+  }) : super(key: key);
   final String title;
-  final Repository repository;
-  const HomePage({Key key, this.title, this.repository}) : super(key: key);
 
   @override
   _HomePageState createState() => _HomePageState();
@@ -17,16 +22,19 @@ class HomePage extends StatefulWidget {
 
 class _HomePageState extends State<HomePage>
     with SingleTickerProviderStateMixin {
-  TabController _tabController;
-  Repository get repository => widget.repository;
+  late TabController _tabController;
 
   @override
   void initState() {
     super.initState();
+    final notifier = context.read<DataNotifier>();
     _tabController = TabController(
-        vsync: this, length: 2, initialIndex: repository.currentIndexTab);
+      vsync: this,
+      length: 4,
+      initialIndex: notifier.currentIndexTab,
+    );
     _tabController.addListener(() {
-      repository.currentIndexTab = _tabController.index;
+      notifier.currentIndexTab = _tabController.index;
     });
   }
 
@@ -43,30 +51,17 @@ class _HomePageState extends State<HomePage>
         title: Text(widget.title),
         actions: <Widget>[
           PopupMenuButton<int>(
-            offset: Offset(0, 40),
+            offset: const Offset(0, 40),
             onSelected: (int value) async {
               switch (value) {
                 case 0:
-                  Navigator.of(context).push(MaterialPageRoute(
-                      builder: (BuildContext context) =>
-                          AvalancheMassifListPage()));
-                  break;
-                case 1:
                   _openAboutDialog(context);
                   break;
               }
             },
             itemBuilder: (BuildContext context) => <PopupMenuItem<int>>[
-              // TODO Fixme Adresse BREA plus valide
-              // PopupMenuItem(
-              //   value: 0,
-              //   child: ListTile(
-              //     leading: Icon(Icons.ac_unit),
-              //     title: Text('Bulletin Avalanche'),
-              //   ),
-              // ),
-              PopupMenuItem(
-                value: 1,
+              const PopupMenuItem(
+                value: 0,
                 child: ListTile(
                   leading: Icon(Icons.info_outline),
                   title: Text('A propos...'),
@@ -76,55 +71,63 @@ class _HomePageState extends State<HomePage>
           ),
         ],
         bottom: TabBar(
-          key: PageStorageKey("tab_key"),
+          key: const PageStorageKey('tab_key'),
           controller: _tabController,
-          tabs: [
-            Tab(text: "Liste", icon: Icon(Icons.list)),
-            Tab(text: "Carte", icon: Icon(Icons.map)),
+          isScrollable: true,
+          tabs: const [
+            Tab(text: 'Stations', icon: Icon(Icons.list)),
+            Tab(text: 'Carte', icon: Icon(Icons.map)),
+            Tab(text: 'Avalanches', icon: Icon(Icons.rss_feed)),
+            Tab(text: 'BREA', icon: Icon(Icons.ac_unit)),
           ],
         ),
       ),
       body: TabBarView(
-        key: PageStorageKey("tab_key"),
+        key: const PageStorageKey('tab_key'),
         controller: _tabController,
-        physics: NeverScrollableScrollPhysics(),
-        children: [
-          ListStationWidget(repository),
-          MapWidget(repository),
+        physics: const NeverScrollableScrollPhysics(),
+        children: const [
+          ListStationWidget(),
+          MapWidget(),
+          AvalancheListWidget(),
+          BREAMassifListView(),
         ],
       ),
     );
   }
 
-  void _openAboutDialog(BuildContext context) {
-    showDialog(
+  Future<void> _openAboutDialog(BuildContext context) async {
+    final packageInfo = await PackageInfo.fromPlatform();
+
+    showDialog<AboutDialog>(
       context: context,
       builder: (BuildContext context) {
         return AboutDialog(
-          applicationName: "Info Neige",
+          applicationName: 'Info Neige',
           applicationVersion:
-              "version: ${repository.packageInfo.version}+${repository.packageInfo.buildNumber}",
+              'version: ${packageInfo.version}+${packageInfo.buildNumber}',
           applicationIcon: Image.asset(
-            "assets/icon/icon.png",
+            'assets/icon/icon.png',
             width: 42,
             height: 42,
           ),
-          applicationLegalese: "MIT",
+          applicationLegalese: 'MIT',
           children: <Widget>[
-            Padding(padding: EdgeInsets.all(5)),
+            const Padding(padding: EdgeInsets.all(5)),
             Text(
               'Développeur: Chonli',
-              style: TextStyle(color: Theme.of(context).textTheme.body1.color),
+              style: TextStyle(
+                  color: Theme.of(context).textTheme.bodyText2?.color),
             ),
-            Padding(padding: EdgeInsets.all(5)),
+            const Padding(padding: EdgeInsets.all(5)),
             InkWell(
-              child: new Text(
+              onTap: () =>
+                  url.launch('https://github.com/Chonli/snow_weather_info'),
+              child: Text(
                 'Lien vers le projet',
-                style:
-                    TextStyle(color: Theme.of(context).textTheme.body2.color),
+                style: TextStyle(
+                    color: Theme.of(context).textTheme.bodyText1?.color),
               ),
-              onTap: () async =>
-                  await launch('https://github.com/Chonli/snow_weather_info'),
             ),
           ],
         );
