@@ -1,90 +1,92 @@
-import 'package:latlong/latlong.dart';
+import 'package:latlong2/latlong.dart';
 import 'package:snow_weather_info/model/avalanche_bulletin.dart';
 import 'package:snow_weather_info/model/station.dart';
 
-class ConstantDatalist {
-  static final List<Nivose> listNivose = [
+abstract class ConstantDatalist {
+  ConstantDatalist._();
+
+  static final listNivose = <Nivose>[
     //alpes nord
     Nivose(
-      "Aiguilles rouges",
+      'Aiguilles rouges',
       LatLng(45.986125, 6.895937),
       2330,
       'AIGRG',
     ),
     Nivose(
-      "Bellecote",
+      'Bellecote',
       LatLng(45.48881, 6.77163),
       3000,
       'BELLE',
     ),
     Nivose(
-      "Le Chevril",
+      'Le Chevril',
       LatLng(45.50179, 6.95604),
       2560,
       'CHEVR',
     ),
     Nivose(
-      "Bonneval",
+      'Bonneval',
       LatLng(45.35175, 7.05132),
       2720,
       'BONNE',
     ),
     Nivose(
-      "Les rochilles",
+      'Les rochilles',
       LatLng(45.08467, 6.47127),
       2450,
       'ROCHI',
     ),
     Nivose(
-      "Allant",
+      'Allant',
       LatLng(45.66345, 6.21195),
       1630,
       'ALLAN',
     ),
     Nivose(
-      "Grande Parei",
+      'Grande Parei',
       LatLng(45.61463, 6.64750),
       2240,
       'PAREI',
     ),
     Nivose(
-      "Col de Porte",
+      'Col de Porte',
       LatLng(45.2952742, 5.7651929),
       1325,
       'COLPO',
     ),
     Nivose(
-      "St Hilaire",
+      'St Hilaire',
       LatLng(45.31449, 5.86462),
       1700,
       'STHIL',
     ),
     Nivose(
-      "Aigleton",
+      'Aigleton',
       LatLng(45.2311598, 6.0381140),
       2240,
       'AIGLE',
     ),
     Nivose(
-      "Le Gua",
+      'Le Gua',
       LatLng(45.01517, 5.58951),
       1600,
       'LEGUA',
     ),
     Nivose(
-      "Les Ecrins",
+      'Les Ecrins',
       LatLng(44.93681, 6.34602),
       2940,
       'ECRIN',
     ),
     Nivose(
-      "La Meije",
+      'La Meije',
       LatLng(45.01255, 6.26485),
       3100,
       'MEIJE',
     ),
     Nivose(
-      "Galibier",
+      'Galibier',
       LatLng(45.05733, 6.37739),
       2559,
       'GALIB',
@@ -92,31 +94,31 @@ class ConstantDatalist {
     //alpes sud
     ,
     Nivose(
-      "Orcières",
+      'Orcières',
       LatLng(44.71857, 6.33378),
       2294,
       'ORCIE',
     ),
     Nivose(
-      "Col Agnel",
+      'Col Agnel',
       LatLng(44.68873, 6.97625),
       2630,
       'AGNEL',
     ),
     Nivose(
-      "Restefond",
+      'Restefond',
       LatLng(44.34205, 6.79871),
       2550,
       'RESTE',
     ),
     Nivose(
-      "Millefonts",
+      'Millefonts',
       LatLng(44.11819, 7.19226),
       2430,
       'MILLE',
     ),
     Nivose(
-      "Parpaillon",
+      'Parpaillon',
       LatLng(44.49522, 6.63650),
       2545,
       'PARPA',
@@ -124,13 +126,13 @@ class ConstantDatalist {
     //corse
     ,
     Nivose(
-      "Sponde",
+      'Sponde',
       LatLng(42.32448, 8.88478),
       1980,
       'SPOND',
     ),
     Nivose(
-      "Maniccia",
+      'Maniccia',
       LatLng(42.20837, 9.04888),
       2360,
       'MANIC',
@@ -138,234 +140,234 @@ class ConstantDatalist {
     //pyrennée
     ,
     Nivose(
-      "Maupas",
+      'Maupas',
       LatLng(42.71428, 0.54995),
       2430,
       'MAUPA',
     ),
     Nivose(
-      "Port d`'Aula",
+      'Port d`Aula',
       LatLng(42.76993, 1.11493),
       2140,
       'PAULA',
     ),
     Nivose(
-      "Canigou",
+      'Canigou',
       LatLng(42.53322, 2.46096),
       2150,
       'CANIG',
     ),
     Nivose(
-      "Hospitalet",
+      'Hospitalet',
       LatLng(42.57778, 1.81140),
       2293,
       'HOSPI',
     ),
     Nivose(
-      "Puigmal",
+      'Puigmal',
       LatLng(42.38318, 2.09143),
       2467,
       'PUIGN',
     ),
     Nivose(
-      "Soum Couy",
+      'Soum Couy',
       LatLng(42.96157, -0.71995),
       2150,
       'SOUMC',
     ),
     Nivose(
-      "Lac d'ardiden",
+      'Lac d’ardiden',
       LatLng(42.84828, -0.06471),
       2445,
       'LARDI',
     ),
     Nivose(
-      "Aiguillettes",
+      'Aiguillettes',
       LatLng(42.74692, 0.18128),
       2120,
       'AIGTE',
     )
   ];
 
-  static final listAvalancheBulletin = [
-    AvalancheBulletin(
-      "Chablais",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP01.pdf",
-      Mountain.alpes_nord,
+  static final listAvalancheBulletin = <AvalancheBulletin>[
+    const AvalancheBulletin(
+      'Chablais',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA01.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Aravis",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP02.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Aravis',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA02.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Mont-Blanc",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP03.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Mont-Blanc',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA03.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Bauges",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP04.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Bauges',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA04.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Beaufortain",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP05.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Beaufortain',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA05.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Haute-Tarentaise",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP06.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Haute-Tarentaise',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA06.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Chartreuse",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP07.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Chartreuse',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA07.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Belledonne",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP08.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Belledonne',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA08.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Maurienne",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP09.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Maurienne',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA09.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Vanoise",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP10.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Vanoise',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA10.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Haute-Maurienne",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP11.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Haute-Maurienne',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA11.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Grandes-Rousses",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP12.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Grandes-Rousses',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA12.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Thabor",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP13.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Thabor',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA13.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Vercors",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP14.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Vercors',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA14.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Oisans",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP15.pdf",
-      Mountain.alpes_nord,
+    const AvalancheBulletin(
+      'Oisans',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA15.xml',
+      Mountain.alpesNord,
     ),
-    AvalancheBulletin(
-      "Pelvoux",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP16.pdf",
-      Mountain.alpes_sud,
+    const AvalancheBulletin(
+      'Pelvoux',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA16.xml',
+      Mountain.alpesSud,
     ),
-    AvalancheBulletin(
-      "Queyras",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP17.pdf",
-      Mountain.alpes_sud,
+    const AvalancheBulletin(
+      'Queyras',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA17.xml',
+      Mountain.alpesSud,
     ),
-    AvalancheBulletin(
-      "Devoluy",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP18.pdf",
-      Mountain.alpes_sud,
+    const AvalancheBulletin(
+      'Devoluy',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA18.xml',
+      Mountain.alpesSud,
     ),
-    AvalancheBulletin(
-      "Champsaur",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP19.pdf",
-      Mountain.alpes_sud,
+    const AvalancheBulletin(
+      'Champsaur',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA19.xml',
+      Mountain.alpesSud,
     ),
-    AvalancheBulletin(
-      "Embrunnais-Parpaillon",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP20.pdf",
-      Mountain.alpes_sud,
+    const AvalancheBulletin(
+      'Embrunnais-Parpaillon',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA20.xml',
+      Mountain.alpesSud,
     ),
-    AvalancheBulletin(
-      "Ubaye",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP21.pdf",
-      Mountain.alpes_sud,
+    const AvalancheBulletin(
+      'Ubaye',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA21.xml',
+      Mountain.alpesSud,
     ),
-    AvalancheBulletin(
-      "Haut-Var-Haut-Verdon",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP22.pdf",
-      Mountain.alpes_sud,
+    const AvalancheBulletin(
+      'Haut-Var-Haut-Verdon',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA22.xml',
+      Mountain.alpesSud,
     ),
-    AvalancheBulletin(
-      "Mercantour",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP23.pdf",
-      Mountain.alpes_sud,
+    const AvalancheBulletin(
+      'Mercantour',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA23.xml',
+      Mountain.alpesSud,
     ),
-    AvalancheBulletin(
-      "Cinto-Rotondo",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP40.pdf",
+    const AvalancheBulletin(
+      'Cinto-Rotondo',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA40.xml',
       Mountain.corse,
     ),
-    AvalancheBulletin(
-      "Renoso",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP41.pdf",
+    const AvalancheBulletin(
+      'Renoso',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA41.xml',
       Mountain.corse,
     ),
-    AvalancheBulletin(
-      "Pays-Basque",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP64.pdf",
+    const AvalancheBulletin(
+      'Pays-Basque',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA64.xml',
       Mountain.pyrennee,
     ),
-    AvalancheBulletin(
-      "Aspe-Ossau",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP65.pdf",
+    const AvalancheBulletin(
+      'Aspe-Ossau',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA65.xml',
       Mountain.pyrennee,
     ),
-    AvalancheBulletin(
-      "Haute-Bigorre",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP66.pdf",
+    const AvalancheBulletin(
+      'Haute-Bigorre',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA66.xml',
       Mountain.pyrennee,
     ),
-    AvalancheBulletin(
-      "Aure-Louron",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP67.pdf",
+    const AvalancheBulletin(
+      'Aure-Louron',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA67.xml',
       Mountain.pyrennee,
     ),
-    AvalancheBulletin(
-      "Luchonnais",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP68.pdf",
+    const AvalancheBulletin(
+      'Luchonnais',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA68.xml',
       Mountain.pyrennee,
     ),
-    AvalancheBulletin(
-      "Couserans",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP69.pdf",
+    const AvalancheBulletin(
+      'Couserans',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA69.xml',
       Mountain.pyrennee,
     ),
-    AvalancheBulletin(
-      "Haute-Ariege",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP70.pdf",
+    const AvalancheBulletin(
+      'Haute-Ariege',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA70.xml',
       Mountain.pyrennee,
     ),
-    AvalancheBulletin(
-      "Andorre",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP71.pdf",
+    const AvalancheBulletin(
+      'Andorre',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA71.xml',
       Mountain.pyrennee,
     ),
-    AvalancheBulletin(
-      "Orlu-St-Barthelemy",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP72.pdf",
+    const AvalancheBulletin(
+      'Orlu-St-Barthelemy',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA72.xml',
       Mountain.pyrennee,
     ),
-    AvalancheBulletin(
-      "Capcir-Puymorens",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP73.pdf",
+    const AvalancheBulletin(
+      'Capcir-Puymorens',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA73.xml',
       Mountain.pyrennee,
     ),
-    AvalancheBulletin(
-      "Cerdagne-Canigou",
-      "http://www.meteofrance.com/integration/sim-portail/generated/integration/img/produits/pdf/bulletins_bra/OPP74.pdf",
+    const AvalancheBulletin(
+      'Cerdagne-Canigou',
+      'http://api.meteofrance.com/files/mountain/bulletins/BRA74.xml',
       Mountain.pyrennee,
     ),
   ];
