@@ -3,44 +3,48 @@ import 'package:snow_weather_info/model/data_station.dart';
 import 'package:charts_flutter/flutter.dart' as charts;
 
 class DataStationChart extends StatelessWidget {
-  final List<DataStation> _data;
-  const DataStationChart(this._data);
+  const DataStationChart({
+    Key? key,
+    required this.data,
+  }) : super(key: key);
+
+  final List<DataStation> data;
 
   @override
   Widget build(BuildContext context) {
-    List<TimeSeriesData> tsdatasnow = [];
-    List<TimeSeriesData> tsdatanewsnow = [];
-    List<TimeSeriesData> tsdatasnowtemperature = [];
-    List<TimeSeriesDataTemp> tsdatatemperature = [];
-    if (_data != null) {
-      for (var d in _data) {
-        if (d.hasSnowHeight) {
-          tsdatasnow.add(TimeSeriesData(d.date, d.snowHeight * 100));
+    final List<TimeSeriesData> tsdatasnow = [];
+    final List<TimeSeriesData> tsdatanewsnow = [];
+    final List<TimeSeriesData> tsdatasnowtemperature = [];
+    final List<TimeSeriesDataTemp> tsdatatemperature = [];
+    if (data.isNotEmpty) {
+      for (final d in data) {
+        if (d.snowHeight != null) {
+          tsdatasnow.add(TimeSeriesData(d.date, d.snowHeight! * 100));
         }
-        if (d.hasTemperature) {
+        if (d.temperature != null) {
           tsdatatemperature.add(TimeSeriesDataTemp(
             d.date,
-            d.temperature,
-            d.hasTemperatureMin24 ? d.temperatureMin24 : d.temperature,
-            d.hasTemperatureMax24 ? d.temperatureMax24 : d.temperature,
+            d.temperature!,
+            d.temperatureMin24 != null ? d.temperatureMin24! : d.temperature!,
+            d.temperatureMax24 != null ? d.temperatureMax24! : d.temperature!,
           ));
         }
-        if (d.hasTemperatureSnow) {
-          tsdatasnowtemperature.add(TimeSeriesData(d.date, d.temperatureSnow));
+        if (d.temperatureSnow != null) {
+          tsdatasnowtemperature.add(TimeSeriesData(d.date, d.temperatureSnow!));
         }
-        if (d.hasSnowNewHeight) {
-          tsdatanewsnow.add(TimeSeriesData(d.date, d.snowNewHeight * 100));
+        if (d.snowNewHeight != null) {
+          tsdatanewsnow.add(TimeSeriesData(d.date, d.snowNewHeight! * 100));
         }
       }
     } else {
       // Dummy list to prevent data = NULL
-      tsdatasnow.add(TimeSeriesData(DateTime.now(), 0.0));
+      tsdatasnow.add(TimeSeriesData(DateTime.now(), 0));
     }
 
-    var seriesSnow = [
+    final seriesSnow = [
       charts.Series<TimeSeriesData, DateTime>(
         id: 'SnowHeight',
-        displayName: "Neige",
+        displayName: 'Neige',
         colorFn: (_, __) => charts.MaterialPalette.blue.shadeDefault,
         domainFn: (TimeSeriesData data, _) => data.time,
         measureFn: (TimeSeriesData data, _) => data.data,
@@ -48,7 +52,7 @@ class DataStationChart extends StatelessWidget {
       ),
       charts.Series<TimeSeriesData, DateTime>(
         id: 'SnowNewHeight',
-        displayName: "Neige fraîche",
+        displayName: 'Neige fraîche',
         colorFn: (_, __) => charts.MaterialPalette.blue.shadeDefault,
         domainFn: (TimeSeriesData data, _) => data.time,
         measureFn: (TimeSeriesData data, _) => data.data,
@@ -56,10 +60,10 @@ class DataStationChart extends StatelessWidget {
       )..setAttribute(charts.rendererIdKey, 'customBar'),
     ];
 
-    var seriesTemperature = [
+    final seriesTemperature = [
       charts.Series<TimeSeriesDataTemp, DateTime>(
         id: 'Temperature',
-        displayName: "Température",
+        displayName: 'Température',
         colorFn: (_, __) => charts.MaterialPalette.red.shadeDefault,
         domainFn: (TimeSeriesDataTemp data, _) => data.time,
         measureFn: (TimeSeriesDataTemp data, _) => data.temp,
@@ -69,7 +73,7 @@ class DataStationChart extends StatelessWidget {
       ),
       charts.Series<TimeSeriesData, DateTime>(
         id: 'Temperature Snow',
-        displayName: "Température Neige",
+        displayName: 'Température Neige',
         colorFn: (_, __) => charts.MaterialPalette.pink.shadeDefault,
         domainFn: (TimeSeriesData data, _) => data.time,
         measureFn: (TimeSeriesData data, _) => data.data,
@@ -82,19 +86,19 @@ class DataStationChart extends StatelessWidget {
       shrinkWrap: true,
       children: [
         Container(
-          constraints: BoxConstraints(
+          constraints: const BoxConstraints(
             maxHeight: 300,
           ),
           child: charts.TimeSeriesChart(
             seriesSnow,
             animate: true,
-            animationDuration: Duration(milliseconds: 800),
+            animationDuration: const Duration(milliseconds: 800),
             behaviors: [
               charts.SeriesLegend(
                 position: charts.BehaviorPosition.bottom,
                 showMeasures: true,
                 horizontalFirst: false,
-                measureFormatter: (num value) {
+                measureFormatter: (value) {
                   return value == null ? '-' : '${value.toStringAsFixed(1)}cm';
                 },
               ),
@@ -104,33 +108,38 @@ class DataStationChart extends StatelessWidget {
                 topMarginSpec: charts.MarginSpec.fixedPixel(10),
                 rightMarginSpec: charts.MarginSpec.fixedPixel(10),
                 bottomMarginSpec: charts.MarginSpec.fixedPixel(10)),
-            domainAxis: charts.DateTimeAxisSpec(
-                tickFormatterSpec: charts.AutoDateTimeTickFormatterSpec(
-                    day: charts.TimeFormatterSpec(
-                        format: 'd', transitionFormat: 'dd/MM'))),
+            domainAxis: const charts.DateTimeAxisSpec(
+              tickFormatterSpec: charts.AutoDateTimeTickFormatterSpec(
+                day: charts.TimeFormatterSpec(
+                  format: 'd',
+                  transitionFormat: 'dd/MM',
+                ),
+              ),
+            ),
             customSeriesRenderers: [
               charts.BarRendererConfig(
-                  // ID used to link series to this renderer.
-                  customRendererId: 'customBar',
-                  cornerStrategy: const charts.ConstCornerStrategy(30),
-                  fillPattern: charts.FillPatternType.forwardHatch)
+                // ID used to link series to this renderer.
+                customRendererId: 'customBar',
+                cornerStrategy: const charts.ConstCornerStrategy(30),
+                fillPattern: charts.FillPatternType.forwardHatch,
+              )
             ],
           ),
         ),
         Container(
-          constraints: BoxConstraints(
+          constraints: const BoxConstraints(
             maxHeight: 300,
           ),
           child: charts.TimeSeriesChart(
             seriesTemperature,
             animate: true,
-            animationDuration: Duration(milliseconds: 800),
+            animationDuration: const Duration(milliseconds: 800),
             behaviors: [
               charts.SeriesLegend(
                 position: charts.BehaviorPosition.bottom,
                 showMeasures: true,
                 horizontalFirst: false,
-                measureFormatter: (num value) {
+                measureFormatter: (value) {
                   return value == null ? '-' : '${value.toStringAsFixed(1)}°C';
                 },
               ),
@@ -140,10 +149,14 @@ class DataStationChart extends StatelessWidget {
                 topMarginSpec: charts.MarginSpec.fixedPixel(10),
                 rightMarginSpec: charts.MarginSpec.fixedPixel(10),
                 bottomMarginSpec: charts.MarginSpec.fixedPixel(10)),
-            domainAxis: charts.DateTimeAxisSpec(
-                tickFormatterSpec: charts.AutoDateTimeTickFormatterSpec(
-                    day: charts.TimeFormatterSpec(
-                        format: 'd', transitionFormat: 'dd/MM'))),
+            domainAxis: const charts.DateTimeAxisSpec(
+              tickFormatterSpec: charts.AutoDateTimeTickFormatterSpec(
+                day: charts.TimeFormatterSpec(
+                  format: 'd',
+                  transitionFormat: 'dd/MM',
+                ),
+              ),
+            ),
           ),
         ),
       ],
@@ -152,15 +165,17 @@ class DataStationChart extends StatelessWidget {
 }
 
 class TimeSeriesData {
+  TimeSeriesData(this.time, this.data);
+
   final DateTime time;
   final double data;
-  TimeSeriesData(this.time, this.data);
 }
 
 class TimeSeriesDataTemp {
+  TimeSeriesDataTemp(this.time, this.temp, this.tempMin, this.tempMax);
+
   final DateTime time;
   final double temp;
   final double tempMin;
   final double tempMax;
-  TimeSeriesDataTemp(this.time, this.temp, this.tempMin, this.tempMax);
 }
