@@ -1,40 +1,41 @@
 import 'package:alphabet_list_scroll_view/alphabet_list_scroll_view.dart';
 import 'package:flutter/material.dart';
-import 'package:snow_weather_info/data/repository.dart';
+import 'package:provider/provider.dart';
+import 'package:snow_weather_info/data/data_notifier.dart';
 import 'package:snow_weather_info/model/station.dart';
 import 'package:snow_weather_info/ui/station_card.dart';
 
 class ListStationWidget extends StatefulWidget {
-  final Repository _repository;
-  const ListStationWidget(this._repository);
+  const ListStationWidget();
 
   @override
   _ListStationWidgetState createState() => _ListStationWidgetState();
 }
 
 class _ListStationWidgetState extends State<ListStationWidget> {
-  TextEditingController _editingController = TextEditingController();
-  List<AbstractStation> _listStation = List<AbstractStation>();
-  List<AbstractStation> _listFavoriteStation = List<AbstractStation>();
-  Repository get _repository => widget._repository;
+  final _editingController = TextEditingController();
+  final _listStation = <AbstractStation>[];
+  final _listFavoriteStation = <AbstractStation>[];
 
   @override
   void initState() {
-    _listStation.addAll(_repository.nivoses);
-    _listStation.addAll(_repository.stations);
-    _listFavoriteStation.addAll(_repository.favoritesStations);
-    _listStation.sort((a, b) => a.name.compareTo(b.name));
     super.initState();
+    final notifier = context.read<DataNotifier>();
+    _listStation.addAll(notifier.nivoses);
+    _listStation.addAll(notifier.stations);
+    _listFavoriteStation.addAll(notifier.favoritesStations);
+    _listStation.sort((a, b) => a.name.compareTo(b.name));
   }
 
-  void _filterSearchResults(String query) {
-    var dummySearchStationList = _repository.stations;
-    var dummySearchNivoseList = _repository.nivoses;
-    var dummySearchFavoriteList = _repository.favoritesStations;
+  void _filterSearchResults(BuildContext context, String query) {
+    final notifier = context.read<DataNotifier>();
+    final dummySearchStationList = notifier.stations;
+    final dummySearchNivoseList = notifier.nivoses;
+    final dummySearchFavoriteList = notifier.favoritesStations;
     _listStation.clear();
     _listFavoriteStation.clear();
     if (query.isNotEmpty) {
-      List<AbstractStation> dummyListData = List<AbstractStation>();
+      final dummyListData = <AbstractStation>[];
       dummySearchStationList.forEach((item) {
         if (item.name.toLowerCase().contains(query.toLowerCase())) {
           dummyListData.add(item);
@@ -63,6 +64,7 @@ class _ListStationWidgetState extends State<ListStationWidget> {
 
   @override
   Widget build(BuildContext context) {
+    final notifier = context.watch<DataNotifier>();
     return AlphabetListScrollView(
       strList: _listStation.map((d) => d.name).toList(),
       highlightTextStyle: TextStyle(
@@ -70,31 +72,35 @@ class _ListStationWidgetState extends State<ListStationWidget> {
       ),
       showPreview: true,
       itemBuilder: (context, index) {
-        return StationCard(_listStation[index]);
+        return StationCard(station: _listStation[index]);
       },
       indexedHeight: (i) {
         return 106;
       },
       headerWidgetList: <AlphabetScrollListHeader>[
         AlphabetScrollListHeader(
-          icon: Icon(Icons.search),
+          icon: const Icon(Icons.search),
           indexedHeaderHeight: (index) => 86,
           widgetList: [
             Padding(
-              padding: EdgeInsets.only(
-                  left: 8.0, bottom: 8.0, top: 16.0, right: 64.0),
+              padding: const EdgeInsets.only(
+                left: 8,
+                bottom: 8,
+                top: 16,
+                right: 64,
+              ),
               child: TextField(
                 onChanged: (value) {
-                  _filterSearchResults(value);
+                  _filterSearchResults(context, value);
                 },
                 controller: _editingController,
-                decoration: InputDecoration(
-                  labelText: "Recherche",
-                  hintText: "Recherche",
+                decoration: const InputDecoration(
+                  labelText: 'Recherche',
+                  hintText: 'Recherche',
                   prefixIcon: Icon(Icons.search),
                   border: OutlineInputBorder(
                     borderRadius: BorderRadius.all(
-                      Radius.circular(20.0),
+                      Radius.circular(20),
                     ),
                   ),
                 ),
@@ -103,14 +109,15 @@ class _ListStationWidgetState extends State<ListStationWidget> {
           ],
         ),
         AlphabetScrollListHeader(
-            widgetList: _listStation
-                .where((s) => _repository.isFavorite(s))
-                .map((s) => StationCard(s))
-                .toList(),
-            icon: Icon(Icons.star),
-            indexedHeaderHeight: (index) {
-              return 106;
-            }),
+          widgetList: _listStation
+              .where((s) => notifier.isFavorite(s))
+              .map((s) => StationCard(station: s))
+              .toList(),
+          icon: const Icon(Icons.star),
+          indexedHeaderHeight: (index) {
+            return 106;
+          },
+        ),
       ],
     );
   }
